Cleaned up stats code (split elapsed time into new file), fixed bug with time not showing for inspirations.
